abstract StreamWrapper
